+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 ### History
-#### 19/10/25
-- Strategy working
-- few updates at create reservation flow
+
+#### 20/10/25
+- started working at delete/cancel flow strategy
+- started thinking about the implementation of cache for user messages
 - TODO:
+ - startt implementing cache for user messages
+ - finish delete/cancel flow strategy
  - Check types at temporal data flow with ai
  - add all tests for google module
  - all the data that is going to be saved at google sheets, need to be parsed to lowercase
+
+#### 19/10/25
+- Strategy working
+- few updates at create reservation flow
 
 #### 18/10/25
 - Strategy pattern added to control user intentions